backend/requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
fastapi==0.94.0
h11==0.14.0
idna==3.4
motor==3.1.1
Pillow==9.4.0
pydantic==1.10.6
pymongo==4.3.3
requests==2.28.2
sniffio==1.3.0
starlette==0.26.0.post1
typing==3.7.4.3
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.0
